+++ react/MuiCozyTheme/AccordionSummary/AccordionExpandIcon.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
-import Icon from '../Icon'
-import BottomIcon from '../Icons/Bottom'
+import Icon from '../../Icon'
+import BottomIcon from '../../Icons/Bottom'
 
 const AccordionExpandIcon = () => <Icon icon={BottomIcon} width={12} />
