+++ react/MuiCozyTheme/AccordionSummary/index.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import AccordionSummary from '@material-ui/core/AccordionSummary'
 
-import AccordionExpandIcon from '../AccordionExpandIcon'
+import AccordionExpandIcon from './AccordionExpandIcon'
 
 AccordionSummary.defaultProps = {
   expandIcon: <AccordionExpandIcon />
